Moving test_patches to patches/test_functional_api
